update with WaitForScaleToZero in api util
